refactor(CreateProjectForm): ♻️ moved form steps to folder
