+++ components/partials/create/project/steps/ContributorsStep.tsx
@@ -10,7 +10,7 @@ import SearchUsers from "components/SearchUsers";
 import { Agent } from "lib/types";
 import { useFormContext } from "react-hook-form";
 import * as yup from "yup";
-import { CreateProjectValues } from "./CreateProjectForm";
+import { CreateProjectValues } from "../CreateProjectForm";
 
 //
