+++ components/partials/create/project/steps/LocationStep.tsx
@@ -6,7 +6,7 @@ import { isRequired } from "lib/isFieldRequired";
 import { useTranslation } from "next-i18next";
 import { Controller, useFormContext } from "react-hook-form";
 import * as yup from "yup";
-import { CreateProjectValues, ProjectType } from "./CreateProjectForm";
+import { CreateProjectValues, ProjectType } from "../CreateProjectForm";
 
 //
