+++ components/partials/create/project/steps/MainStep.tsx
@@ -8,7 +8,7 @@ import { url } from "lib/regex";
 import { useTranslation } from "next-i18next";
 import { Controller, useFormContext } from "react-hook-form";
 import * as yup from "yup";
-import { CreateProjectValues } from "./CreateProjectForm";
+import { CreateProjectValues } from "../CreateProjectForm";
 
 //
